requirements.txt
aiohttp==3.9.3
asyncpg==0.29.0
dateparser==1.2.0
discord.py==2.4.0a0
distro==1.7.0
fantasy_names==1.0.0
firebase_admin==6.5.0
humanize==4.9.0
lru_dict==1.3.0
moviepy==1.0.3
openai==1.30.3
orjson==3.10.3
Pillow==9.0.1
Pillow==10.3.0
psutil==5.9.8
psutil==5.9.0
pytesseract==0.3.10
pytz==2022.1
redis==5.0.3
Requests==2.32.2
sentry_sdk==1.44.0
setuptools==69.2.0
setuptools==59.6.0
sly==0.5
tomli==2.0.1
uvloop==0.19.0
validators==0.25.0
yarl==1.9.4
